requirements.txt
pandas
kafka-python
redis
pika
minio
streamlit